add loop check for generate
